Added drawers in parent layout
