+++ src/app/(private)/(drawer-routes)/DrawerLayout.tsx
@@ -2,7 +2,6 @@
 import { useSession } from 'next-auth/react';
 import { redirect, useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
-import { useScreenWidth } from '../../ScreenWidthProvider';
 import Logo from '@/assets/Dashboard/waza_logo.svg';
 import Home from '@/assets/Dashboard/home.svg';
 import Barbel from '@/assets/Dashboard/barbell.svg';
@@ -10,41 +9,10 @@ import Apple from '@/assets/Dashboard/apple.svg';
 import Bicep from '@/assets/Dashboard/bicep.svg';
 import World from '@/assets/Dashboard/world.svg';
 import Image from 'next/image';
+import { useScreenWidth } from '../../ScreenWidthProvider';
+import Link from 'next/link';
 
-export function Dashboard() {
-  const router = useRouter();
-  const session = useSession();
-  if (session.data && session.data.user.isNewUser) {
-    redirect('/complete-user');
-  }
-  useEffect(() => {
-    (async () => {
-      let route_type = '';
-      if (!session.data) return;
-      else if (session.data.user.user_type === 'Waza Trainer')
-        route_type = 'waza_trainer';
-      else if (session.data.user.user_type === 'Waza Warrior')
-        route_type = 'waza_warrior';
-      else return;
-      const res = await fetch(
-        `http://localhost:3000/api/${route_type}/?user_id=${session.data.user.user_id}`,
-        {
-          method: 'GET',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        },
-      );
-      if (!res.ok) {
-        const data = await res.json();
-        console.log('User not found:', data);
-        router.push(
-          `/complete-user/${route_type}/${session.data.user.user_id}`,
-        );
-      }
-    })();
-  }, [session, router]);
-
+export function DrawerLayout({ children }: { children: React.ReactNode }) {
   const { screenWidth } = useScreenWidth();
   const isLargeScreen = screenWidth >= 1024;
   const [isLeftDrawerOpen, setIsLeftDrawerOpen] = useState(isLargeScreen);
@@ -81,23 +49,33 @@ export function Dashboard() {
           <ul>
             <li className='flex items-center mb-4'>
               <Image src={Home} alt='Dashboard Icon' width={20} height={20} />
-              <span className='ml-2'>Dashboard</span>
+              <Link href={'dashboard'} className='ml-2'>
+                Dashboard
+              </Link>
             </li>
             <li className='flex items-center mb-4'>
               <Image src={Barbel} alt='Workouts Icon' width={20} height={20} />
-              <span className='ml-2'>My Workouts</span>
+              <Link href={'workouts'} className='ml-2'>
+                My Workouts
+              </Link>
             </li>
             <li className='flex items-center mb-4'>
               <Image src={Apple} alt='Diet Icon' width={20} height={20} />
-              <span className='ml-2'>My Diet</span>
+              <Link href={'diet'} className='ml-2'>
+                My Diet
+              </Link>
             </li>
             <li className='flex items-center mb-4'>
               <Image src={Bicep} alt='Diet Icon' width={20} height={20} />
-              <span className='ml-2'>Training</span>
+              <Link href={'training'} className='ml-2'>
+                Training
+              </Link>
             </li>
             <li className='flex items-center mb-4'>
               <Image src={World} alt='Diet Icon' width={20} height={20} />
-              <span className='ml-2'>Community</span>
+              <Link href={'community'} className='ml-2'>
+                Community
+              </Link>
             </li>
           </ul>
         </nav>
@@ -110,6 +88,7 @@ export function Dashboard() {
         } ${isRightDrawerOpen ? 'lg:mr-64' : 'lg:mr-0'}`}
       >
         {/* Your main content goes here */}
+
         {!isLargeScreen && (
           <>
             <button
@@ -126,6 +105,7 @@ export function Dashboard() {
             </button>
           </>
         )}
+        {children}
       </div>
 
       {/* Right Drawer */}
