+++ src/app/(private)/(drawer-routes)/SessionProvider.tsx
@@ -1,0 +1,3 @@
+'use client';
+
+export { SessionProvider } from 'next-auth/react';
