+++ src/app/(private)/(drawer-routes)/community/page.tsx
@@ -1,0 +1,3 @@
+export default function DashboardPage() {
+  return <div>Community</div>;
+}
