+++ src/app/(private)/(drawer-routes)/diet/page.tsx
@@ -1,0 +1,3 @@
+export default function DashboardPage() {
+  return <div>Diet</div>;
+}
