+++ src/app/(private)/(drawer-routes)/training/page.tsx
@@ -1,0 +1,3 @@
+export default function DashboardPage() {
+  return <div>Training</div>;
+}
