+++ src/app/(private)/(drawer-routes)/workouts/page.tsx
@@ -1,0 +1,3 @@
+export default function DashboardPage() {
+  return <div>Workouts</div>;
+}
